Refactored out common parts using Claude
