+++ .vendor/sse-auth-client/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 .mcp-cli
+dist
