+++ .vendor/sse-auth-client/tsup.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  entry: ['src/client.ts', 'src/proxy.ts'],
+  format: ['esm'],
+  dts: true,
+  clean: true,
+  outDir: 'dist',
+  // external: ['typescript'],
+})
